Enhance terminate command to support granular cleanup options
In an effort to stay consistent with other Genesis options, the `--keep-*` options were replaced with `--[no-]*` options, and the things that could be cleaned up were expanded to include network claims, user-provided secrets and credhub secrets.

By default, everything will be cleaned up, but --no-* options can exclude specific ones.  There is also a `--no-cleanup` option that inverts the default, allowing `--*` (without the no-)

Also expanded the summary of what will change prior to asking the user to confirm (without the -y option), and better details of what would be removed on a dry-run.