Extracted BOSH Director status method
While BOSH::Director::connect_and_validate is a user-facing method, it also runs a bunch of checks on the status of the BOSH director.  By separating the status check from the reporting, we can cleanly call the status code without impacting the output the user sees, while also separating the concerns between determining and reporting.