Added the ability to remove files in Blueprint
There may be times where you have added a core list of files in the
blueprint hook, but as you continue to process features, you may want to
remove files that are no longer needed or actively conflict with the
feature. This commit allows you to remove files from the blueprint
hook. This is done by adding a new key to the hook config called
"remove_files", which takes a list of one or more files to remove.