Added spruce_diff function to Genesis
Added a new function to the Genesis module that allows users to
compare two different versions of a file using the `spruce diff` command.
This function takes two arguments, which can be either file paths or
a hash reference containing a 'file', 'object', or 'content' key,
representing the file, Perl object, or yaml string to be compared.
It can also optionally take a 'label' key to specify the temporary file
context for the comparison (defaults to
'spruce_diff-<random-6-digit-number>.yml').

It returns the result of the `spruce diff` command, in a full-color
UTF8-encoded format, and an $is_diff flag that indicates whether there
was a difference between the two versions.