Add constants for exodus time format
Added EXODUS_TIME_FORMAT and EXODUS_TIME_FORMAT_SHORT constants to
standardize timestamp format across the codebase.

Note: This used to be in Genesis::Env, but was moved to Genesis.pm for
better accessibility and to avoid unnecessary usage of passing $env
objects around or importing Genesis::Env in other modules.