require both BOSH_USER and BOSH_PASSWORD to be set
Previously, the code would set $user and $pw if either BOSH_USER or
BOSH_PASSWORD was present in the environment, but only error if both
were missing when the 'require' policy was set. This could lead to
partial credentials being used, causing authentication failures.

Now, both BOSH_USER and BOSH_PASSWORD must be set together for
credentials to be used. If the 'require' policy is set and either is
missing, the code bails out with an error. Users must ensure both
environment variables are set when required, or update scripts and
CI jobs to provide both variables together.