CloudConfig: stablize get_network_security_groups...
The get_network_security_groups method now sorts returned security
groups for deterministic output.

The vm_extension_definition method is aliased as
vm_extension_definition_for_iaas, as this is a more accurate name
for the method's purpose.  Kit authors should use this alias
instead of the original method name, as it will eventually be
deprecated.