Removed GENESIS_TARGET_VAULT from vault env call
While the SAFE_TARGET and VAULT_ADDR are still set, the
GENESIS_TARGET_VAULT should not get changed in the context of the vault
environment.  This method is used to get the environment used to
allow spruce to interact with the vault service this object is
representing, but GENESIS itself may not be using this vault service
for its own secrets management.  This disconnnect is needed to use a
local vault when entombed secrets are used, but callbacks to Genesis
still need to access the real vault to obtain actual values (not
credhub references stored in the local vault).