Add uniq to runtime config request deduplication
Import the uniq function from Genesis and use it to remove
duplicate runtime config requests. This prevents the same
runtime config from being processed multiple times during
deployment operations.

This comes into play mostly when using the '*' or 'all' special
token to enable all of the runtime builds and then add options
to some of them. The uniq function ensures that even if the same
runtime config is requested multiple times, it will only be
included once in the final list of requests.

No user code changes required - this is an internal
improvement to hook processing logic.